pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "robotframework-testdoc"
dynamic = ["version"]
description = "A CLI Tool to generate a Test Documentation for your RobotFramework Test Scripts."
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Marvin Klerx", email = "marvinklerx20@gmail.com" }
]
license = { text = "Apache-2.0" }

classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: 3.14"
]

dependencies = [
    "click",
    "robotframework",
    "jinja2",
    "tomli",
    "pydantic",
    "mkdocs",
    "mkdocs-macros-plugin",
    "mkdocs-material",
    "mkdocs-include-markdown-plugin",
    "pymdown-extensions"
]

[project.optional-dependencies]
dev = [
    "hatch",
    "ruff",
    "pytest",
    "pre-commit",
    "mypy",
    "robotframework-robocop"
]

[project.scripts]
testdoc = "testdoc.cli:main"

[project.urls]
Repository = "https://github.com/MarvKler/robotframework-testdoc"
Issues = "https://github.com/MarvKler/robotframework-testdoc/issues"


[tool.hatch.build.targets.wheel]
require-runtime-dependencies = true
only-include = ["src/testdoc"]
sources = ["src"]
include = [
  "src/testdoc/html/images/*.svg",
  "src/testdoc/default.toml",
  "src/testdoc/html/rendering/mkdocs/*.py",
  "src/testdoc/html/templates/**/*.html",
  "src/testdoc/html/templates/**/*.md",
  "src/testdoc/html/templates/**/*.css",
  "src/testdoc/html/templates/**/*.yml"
]

[tool.hatch.build.targets.sdist]
include = [
  "src/testdoc/html/images/*.svg",
  "src/testdoc/html/templates/**/*.html",
  "src/testdoc/default.toml",
  "README.md",
]

[tool.hatch.version]
path = "src/testdoc/__about__.py"

[tool.hatch.build]
dev-mode-dirs = ["src"]

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/Tables tests}"

[tool.hatch.envs.dev]
dependencies = [
  "ruff",
  "pytest"
]
[tool.hatch.envs.dev.scripts]
lint = "ruff check src/"
atest = "pytest atest/acceptance"
perftest = "pytest atest/performance"
alltest = "pytest atest"


[tool.ruff]
line-length = 150
src = ["src"]
exclude = [
  ".venv",
  ".git",
  "__pycache__",
  "build",
  "dist",
  "bootstrap.py",
  "atest/",
]

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "C90",
  "I",
  "N",
  "B",
  "PYI",
  "PL",
  "PTH",
  "UP",
  "A",
  "C4",
  "DTZ",
  "ISC",
  "ICN",
  "INP",
  "PIE",
  "T20",
  "PYI",
  "PT",
  "RSE",
  "RET",
  "SIM",
  "RUF"
]
ignore = [
  "N999",  # Robot library names stay CamelCase
]
unfixable = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
strict = true
show_error_codes = true
warn_return_any = true
warn_unused_configs = true
warn_unreachable = true
files = ["src"]
exclude = "/(\\.venv|\\.git|__pycache__|build|dist|atest)/"
disable_error_code = ["import-untyped"]
